% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OldestScorerPerYear.R
\name{OldestScorerPerYear}
\alias{OldestScorerPerYear}
\title{A Function to Find Highest Oldest NBA Player by Year,}
\usage{
OldestScorerPerYear(Yearinput)
}
\description{
This function allows you to Find Oldest NBA Player by Year
}
\examples{
OldestScorerPerYear()
}
\keyword{nba,}
\keyword{oldest,}
